mylab/s01_azure_ai_search/requirements.txt
# Azure AI Search and Documents
azure-search-documents>=11.4.0

# Azure AI Projects and Agents
azure-ai-projects>=1.0.0b1

# Azure Identity for authentication
azure-identity>=1.10.0

# Environment variable management
python-dotenv>=1.0.0

# Optional: Additional utilities
# azure-core>=1.24.0
# azure-common>=1.1.28